Fallback to android icon if package has no icon
The applicationInfo.icon value is 0 if the package has no icon. So we
can set the icon to the default android icon there
(ReadInstalledPackagesAsyncTask) instead of catching the
Resources.NotFoundException exception and falling back to the default
android icon like before. Also the default android icon dummy package
can be static.