Remove old startActivityForResult intent passing code
In the previous implementation of the instrumentation dialog
the dialog was an activity which was started with
startActivityForResult. It returned the instrumentation result
which got passed from the MainActivity to the AppListFragment.
There it was used to display the result snackbar. Now we use
a broadcast receiver to receive the instrumentation results which
makes the intent passing code useless.