Backup merged APK, keep latest copy, closes #7
In the new version of the UI, we dropped the setting to activate merged apk backup.
Now the app keeps a copy of the last merged apk and overwrites the previous one.
This way, we make sure to not fill up too much space but still be able to debug efficiently.

The backup location is managed by the system, so when the user deletes the app the apk gets
removed as well (/sdcard/Android/data/saarland.cispa.artist.artistgui/files/).

/sdcard/Android/data/saarland.cispa.artist.artistgui/files